18.03.21 added custom hook useAbout
