+++ src/index.js
@@ -5,7 +5,6 @@ import App from './Components/App/App';
 //todo Сайт должен отображать хорошие показатели на PageSpeed Insights.
 // todo Должны быть описаны meta-теги для представления в социальных сетях, Open Graph. О чем речь
 //todo Страница должна быть опубликована.
-
-// В создании проекта должны использоваться unit-тесты. (Задание со звездочкой, для самостоятельного разбора темы*)
+// todo В создании проекта должны использоваться unit-тесты. (Задание со звездочкой, для самостоятельного разбора темы*)
 
 ReactDOM.render(<App />, document.getElementById('root'));
